requirements.txt
matplotlib==3.10.1
pandas==2.2.3
streamlit==1.43.1
plotly-express==0.4.1